Moved AlertCreator to alert.py
