+++ modules/diagnostics.py
@@ -1,15 +1,13 @@
-from _weakref import proxy
 from typing import List
 
-import commands2
 from commands2 import CommandScheduler
 from commands2.cmd import sequence
 from wpilib import RobotController
 
 from commands.diagnostics.arm import DiagnoseArm
 from commands.diagnostics.claw import DiagnoseClaw
-from commands.diagnostics.intake import DiagnoseIntake
 from commands.diagnostics.diagnoseall import DiagnoseAll
+from commands.diagnostics.intake import DiagnoseIntake
 from commands.diagnostics.utils.setrunningtest import SetRunningTest
 from modules.hardware import HardwareModule
 from ultime.module import Module, ModuleList
@@ -18,18 +16,18 @@ from ultime.module import Module, ModuleList
 class DiagnosticsModule(Module):
     def __init__(self, hardware: HardwareModule, module_list: ModuleList):
         super().__init__()
+        self.components = hardware.subsystems + module_list.modules
+
         self.components_tests = {
             hardware.intake: DiagnoseIntake(hardware.intake),
             hardware.claw: DiagnoseClaw(hardware.claw),
             hardware.arm: DiagnoseArm(hardware.arm, hardware.elevator),
         }
 
-        self._hardware = proxy(hardware)
-        self._module_list = proxy(module_list)
         self._battery_voltage: List[float] = []
         self._is_test = False
         self._run_all_command = DiagnoseAll(
-            self._hardware,
+            hardware,
             [
                 sequence(
                     SetRunningTest(component, True),
@@ -46,7 +44,7 @@ class DiagnosticsModule(Module):
 
     def testInit(self) -> None:
         CommandScheduler.getInstance().enable()
-        for component in self._hardware.subsystems + self._module_list.modules:
+        for component in self.components:
             component.clearAlerts()
         self._run_all_command.schedule()
         self._is_test = True
@@ -61,8 +59,9 @@ class DiagnosticsModule(Module):
 
         def getComponentsNames() -> List[str]:
             return [
-                subsystem.getName()
-                for subsystem in self._hardware.subsystems + self._module_list.modules
+                component.getName()
+                for component in self.components
+                if len(component.registered_alerts) > 1
             ]
 
         def getBatteryVoltage() -> List[float]:
